refactor: 로컬스토리지 관련 함수 분리
