+++ docs/REQUIREMENTS.md
@@ -72,8 +72,8 @@
 - [x] refactor: 음식점 모달 창 렌더링 함수 분리
 - [x] refactor: 음식점 목록 리렌더링 함수 분리
 - [x] refactor: 초기 화면 랜더링 함수 분리
+- [x] refactor: 로컬스토리지 관련 함수 분리
 - [ ] fix: 모바일 환경에서 모달 스크롤 오류 현상 해결
-- [ ] fix: 텍스트 칼라 오류 해결
 - [ ] feat: 음식점 상세 정보 확인
   - 카테고리, 이름, 거리, 설명, 참고 링크 확인
 - [ ] feat: 음식점 삭제
