+++ src/utils/localStorage.ts
@@ -1,0 +1,9 @@
+import Restaurant from '../domain/Restaurant';
+
+export const saveRestaurants = (restaurants: Restaurant[]) => {
+  localStorage.setItem('restaurants', JSON.stringify(restaurants));
+};
+
+export const getRestaurants = () => {
+  return JSON.parse(localStorage.getItem('restaurants') || '[]');
+};
